Add files via upload
Create templated linked list class and templated node class.   Specifically make it a singly linked list with a head pointer only (no tail pointer).

 

The linked list needs to include at least the following functionality:

InsertFront(data) – Add a new node to the front of the list.

InsertBack(data) – Add a new node to the end of the list.

InsertAt(data, index) – Add a new node to the list at the specified (zero-based) index from the head.  The node currently at that index should become the "next" for the node being added.  If the index is beyond the end of the list throw an appropriate error.

RemoveFront() – Remove the node at the front of the list and return the data it contained.

RemoveBack() - Remove the node at the end of the list and return the data it contained.

Front() – Return the data in the first node but do not remove it.

Back() - Return the data in the last node but do not remove it.

Clear() – Remove all nodes.

[] - Overload the [] operator to access the data in the node at the specified (zero based) index if it exists or throw an appropriate exception if the index is not valid.

 

Note: Throw exceptions if the user tries to Remove a node from an empty list.  Clear called on an empty list should not throw an exception

 

Node:

Ensure that the only class that can create a node is the linked list class.

+List Enhancements Portion
Sort(function pointer to comparison function)

The comparison function pointer takes 2 const T& parameters and has an integer return type.  The function pointer will be used to compare the data elements of pairs of nodes. If the function returns 0 it means that the parameter values are equal, if it returns less than zero it means the first parameter is less than second parameter and if it returns greater than zero it means that the first parameter is greater than second parameter. Use this information to sort the list by manipulating the head pointer and the next pointers of the nodes in the list.  To be clear while performing the sort - it may not result in the constructors, destructors or assignment operators of either the Node class or T being called.  It should not error if called on an empty list, a list with a single node or a list with multiple nodes.  You are welcome to use any sorting algorithm/type of sorting that you like.

ForEach(function pointer to method to execute)

The function pointer takes a T& parameter and has an void return type.   This ForEach method executes the function pointer against each node in the list passing the data value for the node.  It should not error if called on an empty list, a list with a single node or a list with multiple nodes. 

ForEach(function pointer to method to execute, U& additionalData)

The function pointer takes a T& parameter and a U& parameter and has an void return type.  This ForEach method executes the function pointer against each node in the list passing the data value for the node and the additionalData parameter.  It should not error if called on an empty list, a list with a single node or a list with multiple nodes. 

Reverse()

Reverses the order of nodes in the list.  Example: If you start list of { 1, 2, 3, 4, 5 } after reversing the list would be { 5, 4, 3, 2, 1 }.  (Small) extra credit if this reordering doesn't result in the constructors, destructors or assignment operators of either the Node class or T being called.

Range based for loop support

Add range based for loop support to your List.